python_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
pdf2image==1.17.0
pytesseract==0.3.10
pillow==10.1.0
werkzeug==3.0.1